refactor: refactor all app
