+++ react-components/src/components/Layout.tsx
@@ -2,28 +2,31 @@ import React from 'react';
 import { Outlet } from 'react-router-dom';
 import { Navigation } from './Navigation';
 
-const getPathPage = () => {
-  const path = window.location.pathname;
-  let page = '';
-  if (path === '/') {
-    page = 'Home';
-  } else if (path === '/about') {
-    page = 'About Us';
-  } else {
-    page = 'Error';
-  }
+export enum Pages {
+  Main = 'Main',
+  AboutUs = 'About Us',
+  Error = 'Error',
+}
 
-  return page;
+const getPage = () => {
+  switch (window.location.pathname) {
+    case '/':
+      return Pages.Main;
+    case '/about':
+      return Pages.AboutUs;
+    default:
+      return Pages.Error;
+  }
 };
 
-export class Layout extends React.Component<object, { page: string }> {
+export class Layout extends React.Component<object, { page: Pages }> {
   constructor(props: object) {
     super(props);
-    this.state = { page: getPathPage() };
+    this.state = { page: getPage() };
     this.handleClick = this.handleClick.bind(this);
   }
 
-  handleClick(page: string) {
+  handleClick(page: Pages) {
     this.setState({ page: page });
   }
   render() {
