--- react-components/src/pages/About/About.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export class About extends React.Component {
-  // componentDidMount() {
-  //   console.log('componentDidMount()' + window.location.pathname);
-  // }
-  render() {
-    return <div>About</div>;
-  }
-}
