+++ react-components/src/pages/AboutUs/AboutUs.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export class AboutUs extends React.Component {
+  render() {
+    return <div>About</div>;
+  }
+}
