+++ react-components/src/pages/Main/SearchBar.tsx
