+++ react-components/src/tests/App.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { describe, expect, test } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 import { MemoryRouter } from 'react-router-dom';
