+++ react-components/src/tests/Card.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { describe, expect, test } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { Card } from './Card';
+import { Card } from '../pages/Main/Card';
 
 describe('Accordion test', () => {
   test('Should show title', () => {
